Adding SharedVAriables at the same time as refactoring them
I refactored them to all extende the SharedVariable class and to give
them all is_class and get_class methods. Also the signal is now
declared in the base class, so we can have it for completion even
when just checking for the base class.

The base class also adds a connect_to and disconnect_from
methods to use the commom signal as an advantage